Refactoring: new WikiRepoManager class (in 'git_manager.py') with git related methods.
