+++ wiki.py
@@ -1,5 +1,4 @@
 import os
-import datetime
 import time
 import re
 import logging
@@ -10,9 +9,9 @@ import knowledge_graph
 from flask import Flask, render_template, request, redirect, url_for, send_from_directory
 from werkzeug.utils import secure_filename
 from random import randint
-from git import Repo, InvalidGitRepositoryError, GitCommandError, NoSuchPathError
 
 from config import get_config
+from git_manager import WikiRepoManager
 
 
 CONFIG = get_config()
@@ -32,150 +31,7 @@ app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 logger = logging.getLogger('werkzeug')
 logger.setLevel(logging.ERROR)
 
-
-def is_git_repo(path: str) -> bool:
-    """
-    Function that determines if the given path is a git repo.
-    :return: True if is a repo, False otherwise.
-    """
-    try:
-        _ = Repo(path).git_dir
-        return True
-    except (InvalidGitRepositoryError, NoSuchPathError):
-        return False
-
-
-def git_repo_init() -> Repo:
-    """
-    Function that initializes the git repo of the Wiki.
-    The repo can be local or remote (it will be cloned), according to the config setting.
-    :return: initialized repo
-    """
-    git_repo = None
-    if is_git_repo(CONFIG["wiki_directory"]):
-        git_repo = git_get_existing_repo()
-        print("existing")
-    else:
-        if CONFIG["remote_url"]:  # if a remote url has been set, clone the repo
-            git_repo = git_clone_remote()
-            print("remote clone")
-        else:
-            git_repo = git_new_local()
-            print("new local")
-
-    # Configure git username and email
-    if git_repo:  # if the repo has been initialized
-        git_repo.config_writer().set_value("user", "name", "wikmd").release()
-        git_repo.config_writer().set_value("user", "email", "wikmd@no-mail.com").release()
-
-    return git_repo
-
-
-def git_get_existing_repo() -> Repo:
-    """
-    Function that gets the existing repo in the wiki_directory.
-    Could be local or remote.
-    :return: Repo
-    """
-    git_repo = None
-    try:
-        app.logger.info("Initializing existing repo ...")
-        git_repo = Repo(CONFIG["wiki_directory"])
-        git_repo.git.checkout()  # don't specify branch, it could be main or master
-    except (InvalidGitRepositoryError, GitCommandError, NoSuchPathError) as e:
-        app.logger.error(f"Error during existing git repo initialization: {str(e)}")
-
-    return git_repo
-
-
-def git_clone_remote() -> Repo:
-    """
-    Function that clones a remote git repo according to the remote_url and the wiki_directory set in the config.
-    :return: Repo
-    """
-    git_repo = None
-    try:
-        app.logger.info(f"Cloning {CONFIG['remote_url']} ...")
-        git_repo = Repo.clone_from(url=CONFIG["remote_url"], to_path=CONFIG["wiki_directory"])
-    except (InvalidGitRepositoryError, GitCommandError, NoSuchPathError) as e:
-        app.logger.error(f"Error during remote git repo cloning: {str(e)}")
-    return git_repo
-
-
-def git_new_local() -> Repo:
-    """
-    Function that inits a new local git repo according to the wiki_directory set in the config.
-    :return: Repo
-    """
-    git_repo = None
-    try:
-        app.logger.info("Creating a new local git repo ...")
-        git_repo = Repo.init(path=CONFIG["wiki_directory"])
-        git_repo.git.checkout("-b", "main")  # create a new (-b) main branch
-    except (InvalidGitRepositoryError, GitCommandError, NoSuchPathError) as e:
-        app.logger.error(f"Error during new local git repo initialization: {str(e)}")
-    return git_repo
-
-
-def git_pull():
-    """
-    Function that pulls from the remote wiki repo.
-    """
-    try:
-        # git pull
-        repo.git.pull()
-        app.logger.info(f"Pulled from the repo")
-    except Exception as e:
-        app.logger.info(f"Error during git pull: {str(e)}")
-
-
-def git_commit(page_name="", commit_type=""):
-    """
-    Function that commits changes to the wiki repo.
-    :param commit_type: could be 'Add', 'Edit' or 'Remove'.
-    :param page_name: name of the page that has been changed.
-    """
-    try:
-        # git add --all
-        repo.git.add("--all")
-        date = datetime.datetime.now()
-        commit_msg = f"{commit_type} page '{page_name}' on {str(date)}"
-        # git commit -m
-        repo.git.commit('-m', commit_msg)
-        app.logger.info(f"New git commit: {commit_msg}")
-    except Exception as e:
-        app.logger.info(f"Nothing to commit: {str(e)}")
-
-
-def git_push():
-    """
-    Function that pushes changes to the remote wiki repo.
-    """
-    try:
-        # git push
-        repo.git.push("-u", "origin", "master")
-        app.logger.info("Pushed to the repo.")
-    except Exception as e:
-        app.logger.info(f"Error during git push: {str(e)}")
-
-
-def git_sync(page_name="", commit_type=""):
-    """
-    Function that manages the synchronization with a git repo, that could be local or remote.
-    If SYNC_WITH_REMOTE is set, it also pull before committing and then pushes changes to the remote repo.
-    :param commit_type: could be 'Add', 'Edit' or 'Remove'.
-    :param page_name: name of the page that has been changed.
-    """
-    if CONFIG["sync_with_remote"]:
-        git_pull()
-
-    git_commit(page_name=page_name, commit_type=commit_type)
-
-    if CONFIG["sync_with_remote"]:
-        git_push()
-
-
-repo = git_repo_init()
+wrm = WikiRepoManager(flask_app=app)
 
 
 def save(page_name):
@@ -357,7 +213,7 @@ def add_new():
     if request.method == 'POST':
         page_name = fetch_page_name()
         save(page_name)
-        git_sync(page_name, "Add")
+        wrm.git_sync(page_name, "Add")
 
         return redirect(url_for("file_page", file_page=page_name))
     else:
@@ -369,7 +225,7 @@ def edit_homepage():
     if request.method == 'POST':
         page_name = fetch_page_name()
         save(page_name)
-        git_sync(page_name, "Edit")
+        wrm.git_sync(page_name, "Edit")
 
         return redirect(url_for("file_page", file_page=page_name))
     else:
@@ -383,7 +239,7 @@ def edit_homepage():
 def remove(page):
     filename = os.path.join(CONFIG["wiki_directory"], page + '.md')
     os.remove(filename)
-    git_sync(page_name=page, commit_type="Remove")
+    wrm.git_sync(page_name=page, commit_type="Remove")
     return redirect("/")
 
 
@@ -396,7 +252,7 @@ def edit(page):
             os.remove(filename)
 
         save(page_name)
-        git_sync(page_name, "Edit")
+        wrm.git_sync(page_name, "Edit")
 
         return redirect(url_for("file_page", file_page=page_name))
     else:
